Read log tests are working
