+++ test/testreadlog.jl
@@ -1,172 +1,77 @@
 # testing ccall of LCM log file handling
 
-using LCMCore
-importall LCMCore
-
-# also see JuliaLang's RawFD type, but little information found.
-# www.c4learn/c-programming/c-file-structure-and-file-pointer
-mutable struct FILE
-  level::Cshort
-  token::Cshort
-  bsize::Cshort
-  fd::Cuchar
-  flags::Cuint  # (unsigned flags ;)
-  hold::Cuchar
-  buffer::Ptr{Cuchar}
-  curp::Ptr{Cuchar}
-  istemp::Cuint # (unsigned is temp)
-end
-
-mutable struct lcm_eventlog_t
-  f::Ptr{FILE}
-  eventcount::Clonglong
-end
-const _LCMlog = Ptr{lcm_eventlog_t}
-
-
-mutable struct lcm_eventlog_event_t
-  eventnum::Clonglong
-  timestamp::Clonglong
-  channellen::Cint
-  datalen::Cint
-  channel::Ptr{Cuchar}
-  data::Ptr{Cuchar}
-end
-
-
-# LCMCore.liblcm
-
-# channel = "TEST"
+# using Base: Test
+# using LCMCore
 #
+# import LCMCore: encode, decode
+#
+# mutable struct MyMessage
+#     field1::Int32
+#     field2::Float64
+# end
+#
+# compare(a::MyMessage, b::MyMessage; tol::Float64=1e-14) = a.field1 == b.field1 && abs(a.field2 - b.field2) < tol
+#
+# function encode(msg::MyMessage)
+#     buf = IOBuffer()
+#     write(buf, hton(msg.field1))
+#     write(buf, hton(msg.field2))
+#     buf.data
+# end
+#
+# function decode(data, msg::Type{MyMessage})
+#     buf = IOBuffer(data)
+#     MyMessage(ntoh(read(buf, Int32)), ntoh(read(buf, Float64)))
+# end
+
+function handledata(channel, msgdata, groundtruth)
+  msg = decode(msgdata, MyMessage)
+  @test compare(groundtruth, msg)
+  nothing
+end
+
+function handletype(channel, msg, groundtruth)
+  @test compare(groundtruth, msg)
+  nothing
+end
+
+function handlefile(lcl; N=1)
+  for i in 1:N
+    handle(lcl) ? nothing : break
+  end
+  nothing
+end
+
+function main()
+  # record a temporary log file
+  lcmlogdir = joinpath(dirname(@__FILE__),"testdata","testlog.lcm")
+
+  # recreate the messages locally for comparison with those in the test log file
+  msg1 = MyMessage(23, 1.234)
+  msg2 = MyMessage(24, 2.345)
+
+  lc = LCMlog(lcmlogdir)
+  subscribe(lc, "CHANNEL_1", (c, d) -> handledata(c, d, msg1) )
+  subscribe(lc, "CHANNEL_2", (c, m) -> handletype(c, m, msg2), MyMessage)
+  # Consume the log file
+  handlefile(lc, N=100)
+  @test true
+  close(lc)
+  nothing
+end
+
+
+
+
+
 # lcm = LCM()
-#
-# data= rand(UInt8, 10)
-#
-# ccall((:lcm_publish, LCMCore.liblcm), Cint, (Ptr{Void}, Ptr{UInt8}, Ptr{UInt8}, Cuint), lcm, channel, data, length(data))
-
-function lcm_eventlog_create(file::S where S <: AbstractString)
-  ccall(
-    (:lcm_eventlog_create, LCMCore.liblcm),
-    _LCMlog, #Ptr{lcm_eventlog_t},
-    (Cstring, Cstring),
-    file, "r"
-  )
-end
-
-
-function lcm_eventlog_destroy(_log::Ptr{lcm_eventlog_t})
-  ccall(
-    (:lcm_eventlog_destroy, LCMCore.liblcm),
-    Void,
-    (Ptr{lcm_eventlog_t}, ),
-    _log
-  )
-end
-
-
-function lcm_eventlog_read_next_event(_log::Ptr{lcm_eventlog_t})
-    _event = ccall(
-      (:lcm_eventlog_read_next_event, LCMCore.liblcm),
-      Ptr{lcm_eventlog_event_t},
-      (Ptr{lcm_eventlog_t},),
-      _log
-    )
-    unsafe_load(_event)
-end
-
-
-# function fetchRecvBuf(event::)
-#
-# end
-
-
-
-
-
-# typeof(log)
-
-# for i in 1:1000
-#
-# end
-
-
-mutable struct LCMlog
-  _log::_LCMlog
-  subscriptions::Dict{AbstractString, LCMCore.SubscriptionOptions}
-  function LCMlog(filename::S) where {S <: AbstractString}
-    try
-      _log = lcm_eventlog_create(filename)
-      return new(_log, Dict{AbstractString, LCMCore.SubscriptionOptions}())
-    catch e
-      warn(e)
-      backtrace()
-      throw("Cannot open the LCM log file at: $filename")
-    end
-    return new() # should never occur
-  end
-end
-
-
-
-function close(lcmlog::LCMlog)
-  try
-    lcm_eventlog_destroy(lcmlog._log)
-  catch e
-    warn(e)
-  end
-end
-
-
-function handle(lcmlog::LCMlog)::Void
-    event = lcm_eventlog_read_next_event(lcmlog._log)
-    # need a convert from lcm_eventlog_event_t to (RecvBuf, channelbytes, chnlen)
-    rb = LCMCore.RecvBuf(event.data, UInt32(event.datalen), event.timestamp, 0)
-    # need a LCMCore.SubscriptionOptions{T}
-    chn = unsafe_string(event.channel, event.channellen)
-    if haskey(lcmlog.subscriptions, chn)
-      opts = lcmlog.subscriptions[chn]
-      # use onresponse similar to regular live LCM traffic
-      LCMCore.onresponse(rb, event.channel, opts)
-    end
-    nothing
-end
-
-
-function subscribe(lcmlog::LCMlog, channel::S, callback::Function) where {S <: AbstractString}
-  opts = LCMCore.LCMCore.SubscriptionOptions(Void, handletest)
-  lcmlog.subscriptions[channel] = opts
-end
-
-
-function handletest(chn, msgdata)
-  @show "got $chn"
-end
-
-
-log = LCMlog("lcmlog.lcm")
-
-
-subscribe(log, "IMU_SIMULATOR", handletest)
-
-
-# to run handle
-for i in 1:1000
-  handle(log)
-end
-
-event
-
-
-
-
-close(log)
-
-
-
-
-
-
-
+# msg1 = MyMessage(23, 1.234)
+# msg2 = MyMessage(24, 2.345)
+# # run `lcm-logger testlog.lcm`
+# publish(lcm, "CHANNEL_1", msg1)
+# publish(lcm, "CHANNEL_2", msg2)
+# # terminate lcm-logger
+# close(lcm)
 
 
 #
